# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/ping":{"get":{"operationId":"AppController_ping","parameters":[],"responses":{"200":{"description":""}},"summary":"v0.0.5 — Reachability endpoint\n\nInfrastructure-only endpoint.\nNot part of application domain.\nWill be removed or disabled post v0.1.x.","tags":["App"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"summary":"v0.0.6 — POST /auth/register. Body parsed into RegisterDto.","tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"summary":"v0.0.6 — POST /auth/login. Body parsed into LoginDto.","tags":["Auth"]}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"summary":"v0.0.6 — POST /users. Body parsed into CreateUserDto.","tags":["Users"]},"get":{"operationId":"UsersController_read","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}},"summary":"v0.0.6 — GET /users.","tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"summary":"v0.0.6 — GET /users/:id. 404 if id not in mock list.","tags":["Users"]},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"summary":"v0.0.6 — PUT /users/:id. 404 if id not found.","tags":["Users"]},"delete":{"operationId":"UsersController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"v0.0.6 — DELETE /users/:id. 404 if id not found.","tags":["Users"]}},"/files":{"post":{"operationId":"FilesController_upload","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadFileDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponseDto"}}}}},"summary":"v0.0.6 — POST /files. Body parsed into UploadFileDto; no file bytes processed.","tags":["Files"]}},"/files/{id}":{"get":{"operationId":"FilesController_download","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileResponseDto"}}}}},"summary":"v0.0.6 — GET /files/:id. Returns file metadata stub only; no stream.","tags":["Files"]},"delete":{"operationId":"FilesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"v0.0.6 — DELETE /files/:id. 404 if id not found.","tags":["Files"]}},"/sharing":{"post":{"operationId":"SharingController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSharingDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharingResponseDto"}}}}},"summary":"v0.0.6 — POST /sharing. Body parsed into CreateSharingDto.","tags":["Sharing"]},"get":{"operationId":"SharingController_read","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SharingResponseDto"}}}}}},"summary":"v0.0.6 — GET /sharing.","tags":["Sharing"]}},"/sharing/{id}":{"get":{"operationId":"SharingController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharingResponseDto"}}}}},"summary":"v0.0.6 — GET /sharing/:id. 404 if id not in mock list.","tags":["Sharing"]},"put":{"operationId":"SharingController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSharingDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SharingResponseDto"}}}}},"summary":"v0.0.6 — PUT /sharing/:id. 404 if id not found.","tags":["Sharing"]},"delete":{"operationId":"SharingController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"v0.0.6 — DELETE /sharing/:id. 404 if id not found.","tags":["Sharing"]}},"/admin":{"post":{"operationId":"AdminController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdminDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminResponseDto"}}}}},"summary":"v0.0.6 — POST /admin. Body parsed into CreateAdminDto by Nest.","tags":["Admin"]},"get":{"operationId":"AdminController_read","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AdminResponseDto"}}}}}},"summary":"v0.0.6 — GET /admin. Returns list; 200 always for stub.","tags":["Admin"]}},"/admin/{id}":{"get":{"operationId":"AdminController_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminResponseDto"}}}}},"summary":"v0.0.6 — GET /admin/:id. 404 if id not in mock list.","tags":["Admin"]},"put":{"operationId":"AdminController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAdminDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminResponseDto"}}}}},"summary":"v0.0.6 — PUT /admin/:id. 404 if id not found.","tags":["Admin"]},"delete":{"operationId":"AdminController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"v0.0.6 — DELETE /admin/:id. 404 if id not found.","tags":["Admin"]}}},"info":{"title":"KC-Project API","description":"v0.0.8 — Backend API with auto-generated OpenAPI spec","version":"0.0.8","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}}},"AuthResponseDto":{"type":"object","properties":{"token":{"type":"string"},"userId":{"type":"string"},"message":{"type":"string"}},"required":["token","userId"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"},"createdAt":{"type":"string"}},"required":["id","createdAt"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}},"UploadFileDto":{"type":"object","properties":{"filename":{"type":"string"}}},"FileResponseDto":{"type":"object","properties":{"id":{"type":"string"},"filename":{"type":"string"},"size":{"type":"number"},"uploadedAt":{"type":"string"}},"required":["id","filename","uploadedAt"]},"CreateSharingDto":{"type":"object","properties":{"fileId":{"type":"string"},"public":{"type":"boolean"},"expiresAt":{"type":"string"}}},"SharingResponseDto":{"type":"object","properties":{"id":{"type":"string"},"fileId":{"type":"string"},"public":{"type":"boolean"},"createdAt":{"type":"string"},"expiresAt":{"type":"string"}},"required":["id","createdAt"]},"UpdateSharingDto":{"type":"object","properties":{"public":{"type":"boolean"},"expiresAt":{"type":"string"}}},"CreateAdminDto":{"type":"object","properties":{"label":{"type":"string"},"role":{"type":"string"}}},"AdminResponseDto":{"type":"object","properties":{"id":{"type":"string"},"label":{"type":"string"},"role":{"type":"string"},"createdAt":{"type":"string"}},"required":["id","createdAt"]},"UpdateAdminDto":{"type":"object","properties":{"label":{"type":"string"},"role":{"type":"string"}}}}}}